# Improve performance: Don't preprocess per test