This should be done only once per file.

I noticed that each test will generate its own preprocessed file. This seems to have been broken a while back already since there is a comment that explicitly says we want to cache the result.

With this change the bats test suite ran 20% faster on my machine.

I'd like to have a regression test in place for this but so far I have no idea how to achieve this in a sensible manner.

- [x] I have reviewed the [Contributor Guidelines][contributor].
- [x] I have reviewed the [Code of Conduct][coc] and agree to abide by it

[contributor]: https://github.com/bats-core/bats-core/blob/master/docs/CONTRIBUTING.md
[coc]:         https://github.com/bats-core/bats-core/blob/master/docs/CODE_OF_CONDUCT.md