@sublimino 2020-08-02
I have been doing rudimentary benchmarking with https://github.com/sharkdp/hyperfine and Google Cloud Build but don't have a regression test suite yet, just using

```bash
for ((i = 0; i < 200; i++)); do
  cat <<EOF@test "a passing test #${i}" {
  true
}
EOF
done | tee test/fixtures/bats/regression.bats >/dev/null

hyperfine 'bin/bats test/fixtures/bats/regression.bats'
```

And mixing in `--jobs`. Not particularly empirical yet!

This makes a noticeable difference using a single file with 200 no-op tests:

```
# this version
hyperfine 'bin/bats test/fixtures/bats/regression.bats'                                                                                       
Benchmark #1: bin/bats test/fixtures/bats/regression.bats                                                       
  Time (mean ± σ):      1.468 s ±  0.011 s    [User: 1.367 s, System: 0.154 s]                                  
  Range (min … max):    1.453 s …  1.483 s    10 runs                                 

# previous commit
hyperfine 'bin/bats test/fixtures/bats/regression.bats'                                                                                      
Benchmark #1: bin/bats test/fixtures/bats/regression.bats                                                       
  Time (mean ± σ):     18.669 s ±  0.194 s    [User: 18.027 s, System: 0.703 s]                                 
  Range (min … max):   18.379 s … 19.006 s    10 runs